protected routing ok, but can leverage by using context

+++ frontend/src/components/admin/computer/ComputerTable.jsx
@@ -1,5 +1,6 @@
 import { createResource, createSignal } from 'solid-js';
 import { action, useNavigate } from "@solidjs/router";
+import Pagination from '../../utils/Pagination.jsx';
 
 const fetchComputer = async (selectedLab) => {
   const response = await fetch(`${import.meta.env.VITE_LOCALHOST_BACKEND_URL}/computers/lab/${selectedLab}`, {
@@ -15,16 +16,19 @@ const fetchComputer = async (selectedLab) => {
 
 
 function ComputerTable(props){
-  const [currentPage, setCurrentPage] = createSignal(1);
-  const itemsPerPage = 10;
-
   const navigate = useNavigate()
   //harus pake arrow function di props.selectedLab supaya solidJS dapat melacak reaktivitas props.selectedLab dan
   //melakukan refetch setiap kali nilainya berubah
   const [computers , {mutate, refetch}] = createResource(() => props.selectedLab, fetchComputer)
 
+  const [paginated, setPaginated] = createSignal({
+    currentPage: 1,
+    items: [],
+    totalPages: 1
+  });
+
   const handleDelete = async (computerId) => {
-    // if (!confirm("Are you sure you want to delete this computer?")) return;
+    if (!confirm("Are you sure you want to delete this computer?")) return;
     console.log("To be deleted is  : " + computerId)
     const response = await fetch(`${import.meta.env.VITE_LOCALHOST_BACKEND_URL}/computers/id/${computerId}`, {
       method: "DELETE",
@@ -36,22 +40,6 @@ function ComputerTable(props){
       refetch()
     }
     console.log(response)
-  }
-
-  const totalPages = () => {
-    return Math.ceil(computers()?.length / itemsPerPage);
-  }
-
-  const paginatedComputers = () => {
-    const startIndex = (currentPage() - 1) * itemsPerPage;
-    const endIndex = startIndex + itemsPerPage;
-    return computers()?.slice(startIndex, endIndex);
-  }
-
-  const handlePageChange = (page) => {
-    if (page >= 1 && page <= totalPages()) {
-      setCurrentPage(page);
-    }
   }
 
   return(
@@ -67,9 +55,9 @@ function ComputerTable(props){
           </tr>
         </thead>
         <tbody>
-          {paginatedComputers()?.map((computer,index) => (
+          {paginated().items?.map((computer,index) => (
             <tr key={computer.id} class="bg-white border-b border-gray-200 hover:bg-gray-50"> 
-              <td class="px-4 py-3 text-left">{(currentPage() - 1) * itemsPerPage + index + 1}</td>
+              <td class="px-4 py-3 text-left">{(paginated().currentPage - 1) * 10 + index + 1}</td>
               <td class="py-3 text-left whitespace-nowrap">{computer.host_name}</td>
               <td class="py-3 text-left whitespace-nowrap">{computer.ip_address}</td>
               <td class="text-center text-sm px-0.5">
@@ -81,16 +69,8 @@ function ComputerTable(props){
         </table>
 
       </div>
-
-      <div class="w-full flex flex-row justify-between">
-        <button onClick={() => handlePageChange(currentPage() - 1)} disabled={currentPage() === 1} class="ml-0.5 cursor-pointer">
-            Previous
-        </button>
-        <span>Page {currentPage()} of {totalPages()}</span>
-        <button onClick={() => handlePageChange(currentPage() + 1)} disabled={currentPage() === totalPages()} class="cursor-pointer">
-            Next
-        </button>
-      </div>
+      
+      <Pagination items={computers()} onPageChange={setPaginated} />
     </div>
   )
 }
